+++ tests/errors_test.py
@@ -7,6 +7,7 @@ from contentful_management.errors import (
     AccessDeniedError,
     UnauthorizedError,
     VersionMismatchError,
+    UnprocessableEntityError,
     RateLimitExceededError,
     ServerError,
     ServiceUnavailableError,
@@ -237,6 +238,44 @@ class ErrorsTest(TestCase):
         self.assertEqual(str(error), message)
         self.assertTrue(isinstance(error, VersionMismatchError))
 
+    def test_unprocessable_entity_error(self):
+        response = MockResponse(422, {
+            "requestId": "044bb23356babe4f11a3f7f1e77c762a",
+            "message":"The resource you sent in the body is invalid.",
+            "details":{
+                "errors":[
+                    {
+                        "name":"taken",
+                        "path":"display_code",
+                        "value":"en-US"
+                    },
+                    {
+                        "name":"fallback locale creates a loop",
+                        "path":"fallback_code",
+                        "value":"en-US"
+                    }
+                ]
+            },
+            "sys":{
+                "type":"Error",
+                "id":"ValidationFailed"
+            }
+        })
+
+        error = get_error(response)
+
+        self.assertEqual(error.status_code, 422)
+        message = "\n".join([
+            "HTTP status code: 422",
+            "Message: The resource you sent in the body is invalid.",
+            "Details: ",
+            "\t* Name: taken - Path: 'display_code' - Value: 'en-US'",
+            "\t* Name: fallback locale creates a loop - Path: 'fallback_code' - Value: 'en-US'",
+            "Request ID: 044bb23356babe4f11a3f7f1e77c762a"
+        ])
+        self.assertEqual(str(error), message)
+        self.assertTrue(isinstance(error, UnprocessableEntityError))
+
     def test_rate_limit_exceeded_error(self):
         response = MockResponse(429, {
             'message': 'Rate Limit Exceeded'
@@ -344,6 +383,7 @@ class ErrorsTest(TestCase):
             403: "The specified token does not have access to the requested resource.",
             404: "The requested resource or endpoint could not be found.",
             409: "Version mismatch error. The version you specified was incorrect. This may be due to someone else editing the content.",
+            422: "The resource you sent in the body is invalid.",
             429: "Rate limit exceeded. Too many requests.",
             500: "Internal server error.",
             502: "The requested space is hibernated.",